split xwayland.cpp into multiple files for different views
